create getMonthlyMetrics and getMonthlyMetrics.test.js
